reduce diff

+++ vms/platformvm/txs/executor/proposal_tx_executor.go
@@ -27,7 +27,7 @@ const (
 	// SyncBound is the synchrony bound used for safe decision making
 	SyncBound = 10 * time.Second
 
-	MaxValidatorWeightFactor = uint64(5)
+	MaxValidatorWeightFactor = 5
 )
 
 var (
